+++ src/systemctl/systemctl-list-unit-files.c
@@ -144,7 +144,7 @@ int verb_list_unit_files(int argc, char *argv[], void *userdata) {
         _cleanup_(sd_bus_message_unrefp) sd_bus_message *reply = NULL;
         _cleanup_free_ UnitFileList *units = NULL;
         _cleanup_hashmap_free_ Hashmap *h = NULL;
-        unsigned c = 0;
+        size_t c = 0;
         const char *state;
         char *path;
         int r;
